+++ html/index.html
@@ -145,7 +145,7 @@
       <x-keyboard :active-notes="activeNotes" v-bind:octave="subtractor.octave" v-bind:subtractor="subtractor"></x-keyboard>
     </div>
     <footer class="footer">
-      v<%= require('../../package.json').version %> :: Created by&nbsp;<a href="https://github.com/jsakas" target="_blank">Jon Sakas</a>
+      v<%= require('../package.json').version %> :: Created by&nbsp;<a href="https://github.com/jsakas" target="_blank">Jon Sakas</a>
     </div>
   </div>
 </body>
