+++ jest.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+    'collectCoverageFrom': [
+      'core/**/*.js',
+      'ui/**/*.js',
+      '!**/index.js',
+    ]
+  };
