+++ ui/keyboard/Keyboard.scss
@@ -1,4 +1,4 @@
-@import '../style/theme.scss';
+@import 'style/theme.scss';
 
 .keyboard {
   display: block;
